Rename and relocate hooks, provider, and types files
This commit primarily consists of renaming several hooks, providers, and types files with a preceding underscore. Files previously residing in the private directory were moved to a new, more general hooks, providers and types directories. Additionally, the version in package.json has been updated from 0.1.0 to 0.1.1 and a `toValidAssistantString` utility function was added to validate the assistant string.