Refactor useLlamaFunctionListener and add useLlamaFunction
Updated the useLlamaFunctionListener hook to allow function names to be passed in an array, added parsing functionality into it and provide selected arguments as a return value. Further, the useLlamaFunction hook is also added which handles the setting and unsetting of the chat parameters. The functionality has been modularized to make the code more readable and maintainable.