Add jsonrepair and refactor streaming in LlamaStreamingProvider
The dependency jsonrepair was added to enable repairing of JSON objects. The LlamaStreamingProvider.tsx was also refactored; the function beginFunctionStreaming was moved, and the function extendFunctionArguments was extended to handle potentially broken JSON arguments. A utility function removeKeys was also added to facilitate removal of specified keys from an object.